moving validateName to utils
