+++ utils.go
@@ -1,0 +1,22 @@
+package keystore
+
+import (
+	"fmt"
+	"strings"
+)
+
+func validateName(name string) error {
+	if name == "" {
+		return fmt.Errorf("key names must be at least one character")
+	}
+
+	if strings.Contains(name, "/") {
+		return fmt.Errorf("key names may not contain slashes")
+	}
+
+	if strings.HasPrefix(name, ".") {
+		return fmt.Errorf("key names may not begin with a period")
+	}
+
+	return nil
+}
